[feat #32] 게이트웨이 Swagger 설정
