+++ common/dtos/src/main/java/user/LoginRequest.java
@@ -1,0 +1,5 @@
+package user;
+
+public record LoginRequest(String email, String password) {
+
+}
