--- common/dtos/src/main/java/user/UserLookupRequest.java
@@ -1,5 +1,0 @@
-package user;
-
-public record UserLookupRequest(String email, String password) {
-
-}
