+++ docker-compose-redis.yml
@@ -95,3 +95,7 @@ services:
       - redis-replica-1
       - redis-replica-2
       - redis-replica-3
+
+networks:
+  redis:
+    driver: bridge
