+++ gateway/src/main/java/com/ticketPing/gateway/GatewayApplication.java
@@ -1,0 +1,15 @@
+package com.ticketPing.gateway;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.context.annotation.ComponentScan;
+
+@ComponentScan(basePackages = {"com.ticketPing.auth", "caching"})
+@EnableFeignClients
+@SpringBootApplication
+public class GatewayApplication {
+	public static void main(String[] args) {
+		SpringApplication.run(GatewayApplication.class, args);
+	}
+}
