+++ gateway/src/main/java/com/ticketPing/gateway/application/dto/UserCache.java
@@ -1,0 +1,8 @@
+package com.ticketPing.gateway.application.dto;
+
+import java.util.UUID;
+
+public record UserCache(
+        UUID userId,
+        String role
+) { }
