+++ gateway/src/main/java/com/ticketPing/gateway/application/service/QueueCheckService.java
@@ -1,0 +1,7 @@
+package com.ticketPing.gateway.application.service;
+
+public interface QueueCheckService {
+    boolean areTooManyWaitingUsers(String performanceId);
+    boolean isSoldOut(String performanceId);
+    boolean isExistToken(String userId, String performanceId);
+}
