+++ gateway/src/main/java/com/ticketPing/gateway/domain/repository/QueueCheckRepository.java
@@ -1,0 +1,7 @@
+package com.ticketPing.gateway.domain.repository;
+
+public interface QueueCheckRepository {
+    int getAvailableSeats(String performanceId);
+    long getWaitingUsers(String performanceId);
+    Boolean findWorkingToken(String tokenValue);
+}
