+++ gateway/src/main/java/com/ticketPing/gateway/exception/ApplicationException.java
@@ -1,0 +1,14 @@
+package com.ticketPing.gateway.exception;
+
+import com.ticketPing.gateway.presentation.cases.ErrorCase;
+import lombok.Getter;
+
+@Getter
+public class ApplicationException extends RuntimeException {
+    private final ErrorCase errorCase;
+
+    public ApplicationException(ErrorCase errorCase) {
+        super(errorCase.getMessage());
+        this.errorCase = errorCase;
+    }
+}
