+++ gateway/src/main/java/com/ticketPing/gateway/infrastructure/security/SecurityConfig.java
@@ -1,0 +1,37 @@
+package com.ticketPing.gateway.infrastructure.security;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
+import org.springframework.security.config.annotation.web.reactive.EnableWebFluxSecurity;
+import org.springframework.security.config.web.server.ServerHttpSecurity;
+import org.springframework.security.web.server.SecurityWebFilterChain;
+
+@Configuration
+@RequiredArgsConstructor
+@EnableWebFluxSecurity
+public class SecurityConfig {
+
+    private final JwtFilter jwtFilter;
+
+    @Bean
+    public SecurityWebFilterChain securityWebFilterChain(ServerHttpSecurity http) {
+        return http
+                .csrf(ServerHttpSecurity.CsrfSpec::disable)
+                .formLogin(ServerHttpSecurity.FormLoginSpec::disable)
+                .httpBasic(ServerHttpSecurity.HttpBasicSpec::disable)
+                .securityContextRepository(jwtFilter)
+                .authorizeExchange(exchange -> exchange
+                        .pathMatchers("/api/v1/auth/login").permitAll()
+                        .pathMatchers("/api/v1/users/signup").permitAll()
+                        .pathMatchers(HttpMethod.GET, "/api/v1/performances/**").permitAll()
+                        .pathMatchers(HttpMethod.GET, "/api/v1/schedules/{scheduleId}").permitAll()
+                        .pathMatchers("/actuator/prometheus").permitAll()
+                        .pathMatchers("/services/*/v3/api-docs/**", "/swagger-ui/**", "/v3/api-docs/**", "/webjars/**").permitAll()
+                        .anyExchange().permitAll()
+                )
+                .build();
+    }
+
+}
