+++ gateway/src/main/java/com/ticketPing/gateway/presentation/cases/ErrorCase.java
@@ -1,0 +1,8 @@
+package com.ticketPing.gateway.presentation.cases;
+
+import org.springframework.http.HttpStatus;
+
+public interface ErrorCase {
+    HttpStatus getHttpStatus();
+    String getMessage();
+}
