+++ gateway/src/main/java/com/ticketPing/gateway/presentation/response/CustomErrorResponse.java
@@ -1,0 +1,6 @@
+package com.ticketPing.gateway.presentation.response;
+
+import org.springframework.http.HttpStatus;
+
+public record CustomErrorResponse(HttpStatus status, String message) {
+}
