--- gateway/src/main/java/com/ticketping/gateway/GatewayApplication.java
@@ -1,16 +1,0 @@
-package com.ticketping.gateway;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.context.annotation.ComponentScan;
-
-@EnableFeignClients
-@SpringBootApplication
-public class GatewayApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(GatewayApplication.class, args);
-	}
-
-}
