--- gateway/src/main/java/com/ticketping/gateway/application/dto/UserCache.java
@@ -1,8 +1,0 @@
-package com.ticketping.gateway.application.dto;
-
-import java.util.UUID;
-
-public record UserCache(
-        UUID userId,
-        String role
-) { }
