--- gateway/src/main/java/com/ticketping/gateway/application/service/QueueCheckService.java
@@ -1,7 +1,0 @@
-package com.ticketping.gateway.application.service;
-
-public interface QueueCheckService {
-    boolean areTooManyWaitingUsers(String performanceId);
-    boolean isSoldOut(String performanceId);
-    boolean isExistToken(String userId, String performanceId);
-}
