--- gateway/src/main/java/com/ticketping/gateway/domain/repository/QueueCheckRepository.java
@@ -1,7 +1,0 @@
-package com.ticketping.gateway.domain.repository;
-
-public interface QueueCheckRepository {
-    int getAvailableSeats(String performanceId);
-    long getWaitingUsers(String performanceId);
-    Boolean findWorkingToken(String tokenValue);
-}
