--- gateway/src/main/java/com/ticketping/gateway/infrastructure/enums/RedisKeyPrefix.java
@@ -1,16 +1,0 @@
-package com.ticketping.gateway.infrastructure.enums;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum RedisKeyPrefix {
-
-    WAITING_QUEUE("WaitingQueue:"),
-    AVAILABLE_SEATS("AvailableSeats:"),
-    TOKEN_VALUE("Token:");
-
-    private final String value;
-
-}
