--- gateway/src/main/java/com/ticketping/gateway/presentation/cases/ErrorCase.java
@@ -1,8 +1,0 @@
-package com.ticketping.gateway.presentation.cases;
-
-import org.springframework.http.HttpStatus;
-
-public interface ErrorCase {
-    HttpStatus getHttpStatus();
-    String getMessage();
-}
