--- gateway/src/main/java/com/ticketping/gateway/presentation/response/CustomErrorResponse.java
@@ -1,6 +1,0 @@
-package com.ticketping.gateway.presentation.response;
-
-import org.springframework.http.HttpStatus;
-
-public record CustomErrorResponse(HttpStatus status, String message) {
-}
