+++ gateway/src/main/resources/application.yml
@@ -2,10 +2,9 @@ spring:
   application:
     name: gateway
 
-  data:
-    redis:
-     host: redis
-     port: 6379
+  config:
+    import:
+      - "classpath:application-redis.yml"
 
 server:
   port: 10001
@@ -14,6 +13,7 @@ eureka:
   client:
     service-url:
       defaultZone: http://eureka-server:10000/eureka/
+
 management:
   endpoints:
     web:
