+++ gateway/src/test/java/com/ticketPing/gateway/GatewayApplicationTests.java
@@ -1,0 +1,13 @@
+package com.ticketPing.gateway;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GatewayApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
