+++ services/queue-manage/src/main/java/com/ticketPing/queue_manage/infrastructure/config/SwaggerConfig.java
@@ -1,15 +1,28 @@
 package com.ticketPing.queue_manage.infrastructure.config;
 
-import io.swagger.v3.oas.annotations.OpenAPIDefinition;
-import io.swagger.v3.oas.annotations.info.Info;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import io.swagger.v3.oas.models.servers.Server;
+import java.util.List;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-@OpenAPIDefinition(
-        info = @Info(
-                title = "Queue-Manage Service API",
-                version = "v1"
-        )
-)
 @Configuration
 public class SwaggerConfig {
+
+        @Value("${spring.application.name}")
+        private String applicationName;
+
+        @Bean
+        public OpenAPI customOpenAPI() {
+                return new OpenAPI()
+                        .info(new Info()
+                                .title("Queue-Manage Service API")
+                                .version("1.0"))
+                        .servers(List.of(
+                                new Server().url("/")
+                        ));
+        }
+
 }
