+++ services/queue-manage/src/main/java/com/ticketPing/queue_manage/presentaion/controller/WaitingQueueController.java
@@ -3,6 +3,7 @@ package com.ticketPing.queue_manage.presentaion.controller;
 import com.ticketPing.queue_manage.application.dto.GeneralQueueTokenResponse;
 import com.ticketPing.queue_manage.application.service.WaitingQueueService;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -15,6 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Mono;
 import response.CommonResponse;
 
+@Tag(name = "대기열 API")
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/waiting-queue")
